Initial SMS Sync

+++ app/src/main/java/com/cureius/pocket/feature_transaction/data/data_source/TransactionDao.kt
@@ -13,7 +13,7 @@ interface TransactionDao {
     fun getTransactions(): Flow<List<Transaction>>
 
     @Query("SELECT * FROM `transaction` WHERE id = :id")
-    suspend fun getTransactionById(id: Int): Transaction?
+    suspend fun getTransactionById(id: Long): Transaction?
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insertTransaction(transaction: Transaction)
