+++ app/src/main/java/com/cureius/pocket/feature_transaction/domain/repository/TransactionRepository.kt
@@ -6,7 +6,7 @@ import kotlinx.coroutines.flow.Flow
 interface TransactionRepository {
     fun getTransactions(): Flow<List<Transaction>>
 
-    suspend fun getTransactionById(id: Int): Transaction?
+    suspend fun getTransactionById(id: Long): Transaction?
 
     suspend fun insertTransaction(transaction: Transaction)
